backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.26.0
python-dotenv==1.0.0

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9
neo4j==5.16.0

# Prisma (will install via npm/pip)
prisma==0.11.0

# WebSocket & SSE
websockets==12.0
sse-starlette==1.8.2

# AI & LangChain
langchain==0.1.4
langchain-openai==0.0.5
langchain-anthropic==0.0.1
langgraph==0.0.20
langsmith==0.0.77

# Utilities
python-slugify==8.0.1
email-validator==2.1.0
pydantic-extra-types==2.3.0

# Security
cryptography==42.0.0
